[zify] Generate proofs of positivity of constants
The role of the `zify_saturate` tactic is to augment the goal with
positivity constraints.  For instance,
if 0 <= x and 0 <= y we can derive that 0 <= x * y.
The premisses (0 <= x) and (0<= y) are obtained from the context.

The PR also generates proof of positvity (0 <= x) or (0 < x) for x a constant.